Necessary changes

+++ Providers/Modules/Plugins/PatchManagement/plugin/filter_patch_management.rb
@@ -8,7 +8,7 @@ require_relative 'omslog'
 module Fluent
   class LinuxUpdatesFilter < Filter
 
-    Fluent::Plugin.register_filter('filter_patch_management', self)    
+    Fluent::Plugin.register_filter('filter_patch_management', self)
 
     def configure(conf)
       super
@@ -30,14 +30,19 @@ module Fluent
     end
 
     def filter(tag, time, record)
-      begin
-        File.delete("/tmp/az-update-security.list")
-        OMS::Log.info_once("File deleted successfully. /tmp/az-update-security.list")
-      rescue Exception => e
-        if e.errno!=2 #except no such file or directory issue. Errno: ENOENT = 2
-          raise e
+      # checking if rb_file_path exists. if yes, skip below check.
+      if File.exists?("/var/oms/microsoft/omsagent/tmp/rb_file_path") != True
+        begin
+          File.delete("/tmp/az-update-security.list")
+          OMS::Log.info_once("File deleted successfully. /tmp/az-update-security.list")
+        rescue Exception => e
+          if e.errno!=2 #except no such file or directory issue. Errno: ENOENT = 2
+            raise e
+          end
         end
       end
+
+
 
       xml_string = record['xml']
       OMS::Log.info_once("LinuxUpdates : Filtering xml size=#{xml_string.size}")
